Extracted registration logic from UserService to new RegisterService
