+++ src/main/java/pl/szmidla/chatappbackend/api/RegisterApi.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 import pl.szmidla.chatappbackend.data.dto.UserRequest;
+import pl.szmidla.chatappbackend.service.RegisterService;
 import pl.szmidla.chatappbackend.service.UserService;
 
 import javax.validation.Valid;
@@ -13,21 +14,21 @@ import javax.validation.Valid;
 @RequestMapping("/api/register")
 public class RegisterApi {
 
-    private UserService userService;
+    private RegisterService registerService;
 
     @PostMapping(produces = MediaType.APPLICATION_JSON_VALUE)
     public String registerUser(@RequestBody @Valid UserRequest user) {
-        return userService.registerUser(user);
+        return registerService.registerUser(user);
     }
 
     @GetMapping(value = "/username-exists", produces = MediaType.APPLICATION_JSON_VALUE)
     public boolean usernameExists(@RequestParam String username) {
-        return userService.usernameExists(username);
+        return registerService.usernameExists(username);
     }
 
     @GetMapping(value = "/email-exists", produces = MediaType.APPLICATION_JSON_VALUE)
     public boolean emailExists(@RequestParam String email) {
-        return userService.emailExists(email);
+        return registerService.emailExists(email);
     }
 
 }
