+++ src/main/java/pl/szmidla/chatappbackend/service/UserService.java
@@ -5,10 +5,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import pl.szmidla.chatappbackend.data.User;
-import pl.szmidla.chatappbackend.data.dto.UserRequest;
 import pl.szmidla.chatappbackend.data.dto.UserResponse;
 import pl.szmidla.chatappbackend.exception.ItemNotFoundException;
 import pl.szmidla.chatappbackend.repository.UserRepository;
@@ -24,12 +22,7 @@ import java.util.Map;
 @AllArgsConstructor
 @Slf4j
 public class UserService {
-
-    public static String REGISTER_SUCCESS = "Successfully registered.";
-    public static String REGISTER_EMAIL_TAKEN = "This email is already registered.";
-    public static String REGISTER_USERNAME_TAKEN = "This username is already taken.";
     private UserRepository userRepository;
-    private PasswordEncoder passwordEncoder;
 
 
     public User getUserById(long id) {
@@ -48,22 +41,6 @@ public class UserService {
                 });
     }
 
-    public String registerUser(UserRequest userRequest) {
-        if (userRepository.existsByEmail(userRequest.getEmail())) {
-            throw new IllegalArgumentException(REGISTER_EMAIL_TAKEN);
-        }
-        if (userRepository.existsByUsername(userRequest.getUsername())) {
-            throw new IllegalArgumentException(REGISTER_USERNAME_TAKEN);
-        }
-
-        userRequest.setPassword( passwordEncoder.encode(userRequest.getPassword()) );
-        User user = userRequest.toUser();
-        user.setLastActive(LocalDateTime.now(ZoneOffset.UTC));
-        userRepository.save(user);
-
-        return REGISTER_SUCCESS;
-    }
-
     public Page<UserResponse> getNUsersByPhrase(String phrase, int pageNr, int pageSize) {
         if (pageNr < 0) {
             throw new IllegalArgumentException("pageNr < 0");
@@ -71,14 +48,6 @@ public class UserService {
         Pageable usersPage = PageRequest.of(pageNr, pageSize);
         return userRepository.findAllByUsernameContainingIgnoreCase(phrase, usersPage)
                 .map(UserResponse::fromUser);
-    }
-
-    public boolean usernameExists(String username) {
-        return userRepository.existsByUsername(username);
-    }
-
-    public boolean emailExists(String email) {
-        return userRepository.existsByEmail(email);
     }
 
     @Transactional
